Only use ``BuildContext`` where it's needed
